Indirect caller saving, improvments to register saving scheme, fixes to if statements that return in control paths.

+++ oc/compiler/stack/lightstack.c
@@ -126,7 +126,7 @@ u_int8_t lightstack_is_empty(lightstack_t* stack){
 /**
  * Peek from the top of the stack
 */
-u_int16_t lightstack_peek(lightstack_t* stack){
+u_int32_t lightstack_peek(lightstack_t* stack){
 	if(stack->top_index == 0){
 		return 0;
 	} else {
